+++ src/main/java/bamv/training/microposts/dao/TUserListDao.java
@@ -1,15 +1,17 @@
 package bamv.training.microposts.dao;
 
 import bamv.training.microposts.dto.UserDto;
+import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
+@Repository
 public interface TUserListDao {
 
     //アクセス修飾子をすべて削除
-    List<UserDto> findAll();
+    List<UserDto> findAllUser();
 
-    int follow(String userId, String followeeId);
+    int addNewFollow(String userId, String followeeId);
 
-    int unfollow(String userId, String followeeId);
+    int deleteExistFollow(String userId, String followeeId);
 }
