+++ src/main/java/bamv/training/microposts/dao/impl/TUserListDaoImpl.java
@@ -5,26 +5,26 @@ import bamv.training.microposts.dto.UserDto;
 import bamv.training.microposts.service.impl.SequenceServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.stereotype.Repository;
 import org.springframework.stereotype.Service;
 
 import  java.util.List;
 import java.util.Map;
 
-@Service
+@Repository
 public class TUserListDaoImpl implements TUserListDao {
     @Autowired
     private JdbcTemplate jdbcTemplate;
     @Autowired
     private SequenceServiceImpl sequenceService;
 
-    //これがいらなかったぽい。
     //@Autowired
     //UserListDao(JdbcTemplate jdbcTemplate, SequenceServiceImpl sequenceService) {
     //    this.jdbcTemplate = jdbcTemplate;
     //    this.sequenceService = sequenceService;
     //}
     @Override
-    public List<UserDto> findAll() {
+    public List<UserDto> findAllUser() {
         String query = "SELECT user_id AS userId, name FROM m_user";
 
         List<Map<String, Object>> result = jdbcTemplate.queryForList(query);
@@ -38,7 +38,7 @@ public class TUserListDaoImpl implements TUserListDao {
 
     //INSERTの実行前に同一userId, followeeIdでレコードがあれば何もせずにreturnする
     @Override
-    public int follow(String userId, String followeeId) {
+    public int addNewFollow(String userId, String followeeId) {
         String checkDuplicateSql = "SELECT COUNT(*) FROM t_follow WHERE following_user_id = ? AND followed_user_id = ?";
         int count = jdbcTemplate.queryForObject(checkDuplicateSql, Integer.class, userId, followeeId);
         if (count != 0) //すでにフォローしている
@@ -50,7 +50,7 @@ public class TUserListDaoImpl implements TUserListDao {
     }
 
     @Override
-    public int unfollow(String userId, String followeeId) {
+    public int deleteExistFollow(String userId, String followeeId) {
         int number = jdbcTemplate.update("DELETE FROM t_follow WHERE following_user_id = ? AND followed_user_id = ?", userId, followeeId);
         return number;
     }
